+++ mobile/library/cc/engine_builder.cc
@@ -207,6 +207,17 @@ EngineBuilder& EngineBuilder::addStringAccessor(const std::string& name,
   return *this;
 }
 
+EngineBuilder& EngineBuilder::addNativeFilter(const std::string& name,
+                                              const std::string& typed_config) {
+  native_filter_chain_.emplace_back(name, typed_config);
+  return *this;
+}
+
+EngineBuilder& EngineBuilder::addPlatformFilter(const std::string& name) {
+  platform_filters_.push_back(name);
+  return *this;
+}
+
 std::string EngineBuilder::generateConfigStr() const {
   std::vector<std::pair<std::string, std::string>> replacements {
     {"connect_timeout", fmt::format("{}s", this->connect_timeout_seconds_)},
@@ -277,6 +288,19 @@ std::string EngineBuilder::generateConfigStr() const {
   }
   if (this->enable_http3_) {
     insertCustomFilter(alternate_protocols_cache_filter_insert, config_template);
+  }
+
+  for (const NativeFilterConfig& filter : native_filter_chain_) {
+    std::string filter_config = absl::StrReplaceAll(
+        native_filter_template, {{"{{ native_filter_name }}", filter.name_},
+                                 {"{{ native_filter_typed_config }}", filter.typed_config_}});
+    insertCustomFilter(filter_config, config_template);
+  }
+
+  for (const std::string& name : platform_filters_) {
+    std::string filter_config =
+        absl::StrReplaceAll(platform_filter_template, {{"{{ platform_filter_name }}", name}});
+    insertCustomFilter(filter_config, config_template);
   }
 
   config_builder << config_template;